+++ packages/Chem/src/analysis/molecular-matched-pairs/mmp-viewer.ts
@@ -13,6 +13,8 @@ import {getMmpTrellisPlot} from './mmp-frag-vs-frag';
 import {getMmpScatterPlot, runMmpChemSpace} from './mmp-cliffs';
 import {getGenerations} from './mmp-generations';
 
+
+import {drawMoleculeLabels} from '../../rendering/molecule-label';
 import {ILineSeries, MouseOverLineEvent, ScatterPlotLinesRenderer}
   from '@datagrok-libraries/utils/src/render-lines-on-sp';
 import BitArray from '@datagrok-libraries/utils/src/bit-array';
@@ -85,7 +87,7 @@ export class MatchedMolecularPairsViewer extends DG.JsViewer {
     this.activities = this.stringList('activities');
     this.fragmentCutoff = this.float('fragmentCutoff');
 
-    this.totalData = this.string('totalData', 'null', {userEditable: false});
+    this.totalData = this.string('totalData', 'null', {userEditable: false, includeInLayout: false});
   }
 
   onTableAttached() {
@@ -183,13 +185,13 @@ export class MatchedMolecularPairsViewer extends DG.JsViewer {
   }
 
   setupFilters(mmpFilters: MmpFilters, linesActivityCorrespondance: Uint32Array, tp: DG.Viewer): void {
-    for (let i = 0; i < mmpFilters.activitySliderInputs.length; i ++) {
-      mmpFilters.activityActiveInputs[i].onChanged(() => {
+    for (let i = 0; i < mmpFilters.activitySliderInputs.length; i++) {
+      mmpFilters.activityActiveInputs[i].onChanged.subscribe(() => {
         this.refilterCliffs(mmpFilters.activitySliderInputs.map((si) => si.value),
           mmpFilters.activityActiveInputs.map((ai) => ai.value), true);
       });
 
-      mmpFilters.activitySliderInputs[i].onChanged(() => {
+      mmpFilters.activitySliderInputs[i].onChanged.subscribe(() => {
         mmpFilters.activityValuesDivs[i].innerText = mmpFilters.activitySliderInputs[i].value === 0 ? '0' :
           getSigFigs(mmpFilters.activitySliderInputs[i].value, 4).toString();
         this.refilterCliffs(mmpFilters.activitySliderInputs.map((si) => si.value),
@@ -197,7 +199,7 @@ export class MatchedMolecularPairsViewer extends DG.JsViewer {
       });
 
       mmpFilters.activityColorInputs[i].value = this.colorPalette!.hex[i];
-      mmpFilters.activityColorInputs[i].onChanged(() => {
+      mmpFilters.activityColorInputs[i].onChanged.subscribe(() => {
         const progressRendering = DG.TaskBarProgressIndicator.create(`Changing colors...`);
 
         //refresh lines
@@ -238,9 +240,8 @@ export class MatchedMolecularPairsViewer extends DG.JsViewer {
       this.cutoffMasks![i].setAll(true);
     }
 
-    mmpFilters.pairsSliderInput.onChanged(() => {
-      mmpFilters.pairsValueDiv.innerText = mmpFilters.pairsSliderInput.value.toString();
-      const value = mmpFilters.pairsSliderInput.value;
+    mmpFilters.pairsSliderInput.onChanged.subscribe((value) => {
+      mmpFilters.pairsValueDiv.innerText = value.toString();
 
       this.fragmentsMask!.setAll(false);
 
@@ -481,6 +482,8 @@ export class MatchedMolecularPairsViewer extends DG.JsViewer {
 
     const sp = getMmpScatterPlot(mmpInput, embedColsNames);
 
+    drawMoleculeLabels(mmpInput.table, mmpInput.molecules, sp as DG.ScatterPlotViewer, 20, 7, 100, 110);
+
     //running internal chemspace
     const linesEditor = runMmpChemSpace(mmpInput, sp, lines, linesIdxs, linesActivityCorrespondance,
       transPairsGrid.dataFrame, diffs, module, embedColsNames);
@@ -575,6 +578,11 @@ export class MatchedMolecularPairsViewer extends DG.JsViewer {
       this.transPairsGrid!.setOptions({
         pinnedRowValues: [idxPairs.toString()],
         pinnedRowColumnNames: [MMP_NAMES.PAIRNUM],
+      });
+    } else {
+      this.transPairsGrid!.setOptions({
+        pinnedRowValues: [],
+        pinnedRowColumnNames: [],
       });
     }