+++ packages/ODEs/src/app.ts
@@ -855,10 +855,10 @@ export class DiffStudio {
       //@ts-ignore
       options = getOptions(key, ivp.arg[key], CONTROL_EXPR.ARG);
       const input = ui.input.forProperty(DG.Property.fromOptions(options));
-      input.onChanged(async () => {
-        if (input.value !== null) {
+      input.onChanged.subscribe(async (value) => {
+        if (value !== null) {
           //@ts-ignore
-          ivp.arg[key].value = input.value;
+          ivp.arg[key].value = value;
           await this.solve(ivp, getInputsPath());
         }
       });
@@ -870,9 +870,9 @@ export class DiffStudio {
     ivp.inits.forEach((val, key) => {
       options = getOptions(key, val, CONTROL_EXPR.INITS);
       const input = ui.input.forProperty(DG.Property.fromOptions(options));
-      input.onChanged(async () => {
-        if (input.value !== null) {
-        ivp.inits.get(key)!.value = input.value;
+      input.onChanged.subscribe(async (value) => {
+        if (value !== null) {
+        ivp.inits.get(key)!.value = value;
         await this.solve(ivp, getInputsPath());
         }
       });
@@ -885,9 +885,9 @@ export class DiffStudio {
       ivp.params.forEach((val, key) => {
         options = getOptions(key, val, CONTROL_EXPR.PARAMS);
         const input = ui.input.forProperty(DG.Property.fromOptions(options));
-        input.onChanged(async () => {
-          if (input.value !== null) {
-            ivp.params!.get(key)!.value = input.value;
+        input.onChanged.subscribe(async (value) => {
+          if (value !== null) {
+            ivp.params!.get(key)!.value = value;
             await this.solve(ivp, getInputsPath());
           }
         });
@@ -902,9 +902,9 @@ export class DiffStudio {
       options.inputType = INPUT_TYPE.INT; // since it's an integer
       options.type = DG.TYPE.INT; // since it's an integer
       const input = ui.input.forProperty(DG.Property.fromOptions(options));
-      input.onChanged(async () => {
-        if (input.value !== null) {
-          ivp.loop!.count.value = input.value;
+      input.onChanged.subscribe(async (value) => {
+        if (value !== null) {
+          ivp.loop!.count.value = value;
           await this.solve(ivp, getInputsPath());
         }
       });
@@ -992,7 +992,7 @@ export class DiffStudio {
       units: WARNING.UNITS,
     };
     const maxTimeInput = ui.input.forProperty(DG.Property.fromOptions(opts));
-    maxTimeInput.onInput(() => {
+    maxTimeInput.onInput.subscribe(() => {
       if (maxTimeInput.value >= UI_TIME.SOLV_TIME_MIN_SEC)
         this.secondsLimit = maxTimeInput.value;
       else maxTimeInput.value = this.secondsLimit;