+++ packages/PhyloTreeViewer/src/viewers/inject-tree-to-grid.ts
@@ -80,9 +80,9 @@ export async function injectTreeToGridUI(
 
     treeRoot.appendChild(cutSlider.root);
 
-    subs.push(cutSlider.onChanged(() => {
-      console.debug('PhyloTreeViewer: injectTreeToGrid() cutSlider.onChanged() ' + `${cutSlider!.value}`);
-      th.cutTreeToGrid(newickRoot, cutSlider!.value!, grid.dataFrame, leafColName, 'Cluster');
+    subs.push(cutSlider.onChanged.subscribe((value) => {
+      console.debug('PhyloTreeViewer: injectTreeToGrid() cutSlider.onChanged() ' + `${value}`);
+      th.cutTreeToGrid(newickRoot, value!, grid.dataFrame, leafColName, 'Cluster');
     }));
   }