+++ tools/package-template/webpack.config.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const packageName = path.parse(require('./package.json').name).name.toLowerCase().replace(/-/g, '');
-
+const FuncGeneratorPlugin = require('datagrok-tools/plugins/func-gen-plugin');
 module.exports = {
   cache: {
     type: 'filesystem',
@@ -29,4 +29,7 @@ module.exports = {
     libraryTarget: 'var',
     path: path.resolve(__dirname, 'dist'),
   },
+  plugins: [
+    new FuncGeneratorPlugin({outputPath: './src/package.g.ts'}),
+  ],
 };