+++ tools/CHANGELOG.md
@@ -1,6 +1,29 @@
 # Datagrok-tools changelog
 
 
+## 4.13.20 (2024-08-28)
+
+### Features
+
+* Grok link command fixes 
+## 4.13.19 (2024-08-27)
+
+### Features
+
+* Package ts data insertation fixes
+
+## 4.13.18 (2024-08-27)
+
+### Features
+
+* Creation fixes
+
+## 4.13.17 (2024-08-27)
+
+### Features
+
+* Added decorators for viewer functions
+
 ## 4.13.16 (2024-08-23)
 
 ### Features